requirements.txt
requests
beautifulsoup4
lxml
pillow
boto3
python-dotenv